+++ platform-operator/controllers/verrazzano/component/keycloak/keycloak.go
@@ -6,7 +6,16 @@ package keycloak
 import (
 	"bytes"
 	"context"
+	"encoding/json"
+	"errors"
 	"fmt"
+	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/nginx"
+	vzos "github.com/verrazzano/verrazzano/platform-operator/internal/os"
+	corev1 "k8s.io/api/core/v1"
+	"os/exec"
+	"path/filepath"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"text/template"
 
 	"github.com/verrazzano/verrazzano/pkg/bom"
@@ -44,6 +53,22 @@ const kcInitContainerValueTemplate = `
         - name: cacerts
           mountPath: /cacerts
 `
+
+// KeycloakClients represents an array of clients currently configured in Keycloak
+type KeycloakClients []struct {
+	ID       string `json:"id"`
+	ClientID string `json:"clientId"`
+}
+
+type bashFuncSig func(inArgs ...string) (string, string, error)
+
+var bashFunc bashFuncSig = vzos.RunBash
+
+func setBashFunc(f bashFuncSig) {
+	bashFunc = f
+}
+
+var execCommand = exec.Command
 
 // imageData needed for template rendering
 type imageData struct {
@@ -134,4 +159,79 @@ func getEnvironmentName(envName string) string {
 	}
 
 	return envName
+}
+
+func updateKeycloakUris(ctx spi.ComponentContext) error {
+	var keycloakClients KeycloakClients
+
+	if ctx.EffectiveCR().Spec.Profile != vzapi.ManagedCluster {
+		// Get the Keycloak admin password
+		secret := &corev1.Secret{}
+		err := ctx.Client().Get(context.TODO(), client.ObjectKey{
+			Namespace: "keycloak",
+			Name:      "keycloak-http",
+		}, secret)
+		if err != nil {
+			ctx.Log().Errorf("Keycloak Post Upgrade: Error retrieving Keycloak password: %s", err)
+			return err
+		}
+		pw := secret.Data["password"]
+		keycloakpw := string(pw)
+		if keycloakpw == "" {
+			return errors.New("Keycloak Post Upgrade: Error retrieving Keycloak password")
+		}
+		ctx.Log().Info("Keycloak Post Upgrade: Successfully retrieved Keycloak password")
+
+		// Login to Keycloak
+		cmd := execCommand("kubectl", "exec", "keycloak-0", "-n", "keycloak", "-c", "keycloak", "--",
+			"/opt/jboss/keycloak/bin/kcadm.sh", "config", "credentials", "--server", "http://localhost:8080/auth", "--realm", "master", "--user", "keycloakadmin", "--password", keycloakpw)
+		_, err = cmd.Output()
+		if err != nil {
+			ctx.Log().Errorf("Keycloak Post Upgrade: Error logging into Keycloak: %s", err)
+			return err
+		}
+		ctx.Log().Info("Keycloak Post Upgrade: Successfully logged into Keycloak")
+
+		// Get the Client ID JSON array
+		cmd = execCommand("kubectl", "exec", "keycloak-0", "-n", "keycloak", "-c", "keycloak", "--", "/opt/jboss/keycloak/bin/kcadm.sh", "get", "clients", "-r", "verrazzano-system", "--fields", "id,clientId")
+		out, err := cmd.Output()
+		if err != nil {
+			ctx.Log().Errorf("Keycloak Post Upgrade: Error retrieving ID for client ID, zero length: %s", err)
+			return err
+		}
+		if len(string(out)) == 0 {
+			return errors.New("Keycloak Post Upgrade: error retrieving Clients JSON from Keycloak, zero length")
+		}
+		json.Unmarshal([]byte(out), &keycloakClients)
+
+		// Extract the id associated with ClientID verrazzano-pkce
+		var id = ""
+		for _, client := range keycloakClients {
+			if client.ClientID == "verrazzano-pkce" {
+				id = client.ID
+				ctx.Log().Debugf("Keycloak Clients ID found = %s", id)
+			}
+		}
+		if id == "" {
+			return errors.New("Keycloak Post Upgrade: Error retrieving ID for Keycloak user, zero length")
+		}
+		ctx.Log().Info("Keycloak Post Upgrade: Successfully retrieved clientID")
+
+		// Get DNS Domain Configuration
+		dnsSubDomain, err := nginx.BuildDNSDomain(ctx.Client(), ctx.EffectiveCR())
+		if err != nil {
+			ctx.Log().Errorf("Keycloak Post Upgrade: Error retrieving DNS sub domain: %s", err)
+			return err
+		}
+		ctx.Log().Infof("Keycloak Post Upgrade: DNSDomain returned %s", dnsSubDomain)
+
+		// Call the Script and Update the URIs
+		scriptName := filepath.Join(config.GetInstallDir(), "update-kiali-redirect-uris.sh")
+		if _, stderr, err := bashFunc(scriptName, id, dnsSubDomain); err != nil {
+			ctx.Log().Errorf("Keycloak Post Upgrade: Failed updating KeyCloak URIs %s: %s", err, stderr)
+			return err
+		}
+	}
+	ctx.Log().Info("Keycloak Post Upgrade: Successfully Updated Keycloak URIs")
+	return nil
 }
